requirements.txt
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
decorator==5.1.0
Flask==2.0.1
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
ipython==7.27.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
python-dotenv==0.19.0
requests==2.26.0
SQLAlchemy==1.4.23
traitlets==5.1.0
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
xendit-python==0.1.1
